iox#25: ChunkTuple in types
